finish checkout

+++ Doanlesg/src/main/java/com/example/Doanlesg/model/CartItem.java
@@ -1,6 +1,8 @@
 package com.example.Doanlesg.model;
 
 import jakarta.persistence.*;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -19,7 +21,7 @@ public class CartItem {
     @Column(name = "quantity", nullable = false)
     private int quantity;
     @Column(name = "price_at_addition",nullable = false)
-    private double priceAtAddition;
+    private BigDecimal priceAtAddition;
 
     public Long getId() {
         return id;
@@ -53,11 +55,11 @@ public class CartItem {
         this.quantity = quantity;
     }
 
-    public double getPriceAtAddition() {
+    public BigDecimal getPriceAtAddition() {
         return priceAtAddition;
     }
 
-    public void setPriceAtAddition(double priceAtAddition) {
+    public void setPriceAtAddition(BigDecimal priceAtAddition) {
         this.priceAtAddition = priceAtAddition;
     }
 }
